gfs2-utils: build: Avoid using the kernel versions of kernel headers
Currently the configure script checks for kernel headers at
/lib/modules/`uname -r`/source/include/linux/* which means we're using
kernel headers not meant to be used in userspace code. Thus we see a lot
of "Attempt to use kernel headers from user space" warnings when
building with a kernel-devel package installed.

Instead we should be using the userspace versions of the headers in
/usr/include/linux/. This patch removes the config options and checks
which pull in the kernel includes and lets the configure script find the
linux headers in the default include path.

Also removes an unnecessary include of kdev_t.h and updates the linux
header checks in configure.ac to only check for the ones we use.

Signed-off-by: Andrew Price <anprice@redhat.com>