libgfs2: Add a small testing language UI
This adds a minimal 'gfs2l' command to allow easy use of the testing
language. It takes one -f option to specify a script file and a single
argument to specify the path to the fs.

It can be used to build a rudimentary REPL e.g.

  # while read s; do echo "$s" | gfs2l -f- /dev/foo; done
  get sb
  ...

But it's mainly intended to be used as a component of in-tree test
scripts and will not be installed by the build system.

Signed-off-by: Andrew Price <anprice@redhat.com>