gfs2-utils: Added a new script called gfs2_lockcapture that will capture lockdump data.
The script gfs2_lockcapture will capture lockdump data(including dlm debug data)
for the mounted GFS2 filesystems on a cluster node. In addition to the debug
data, uname -a, hostname, date, mount, ps, etc are gathered. The script contains
many configurable optons which include the number of iterations, sleep time
between runs, etc.  The script is completed including docstrings. The Makefile
was also changed so that a similar script was removed called gfs2_lockgather and
was replaced with gfs2_lockcapture as the item that will be made.

Signed-off-by: Shane Bradley <sbradley@redhat.com>