gfs2-utils: Add the beginnings of a test suite
This adds a skeleton test suite based on the Check testing framework.
The tests are only enabled when the check(-devel) package is present
and the suite is run with `make check'.

The Check framework <http://check.sourceforge.net/> is very
straightforward and conceptually consists of test suites containing test
cases containing tests. To define tests we write functions which
exercise our code and make assertions with functions provided by Check,
such as fail_unless().

To begin with there is only one test and it simply calls
lgfs2_selfcheck, which used to be called in the gfs2_edit code.

Signed-off-by: Andrew Price <anprice@redhat.com>