gfs2_trace: Added a script called gfs2_trace for kernel tracing debugging.
The script gfs2_trace is a tool used for debugging kernel tracing events. The
script can enabled or disable all or selected GFS2 kernel tracing events. The
script can capture the trace events and write them to a specific file which will
be tarred up after the file is written.

Signed-off-by: Shane Bradley <sbradley@redhat.com>