gfs2_lockcapture: The script now returns a correct exit code when the script exits.
If there was files(not directories) created then the script will exit with 0,
else 1 since there was no lockdump files captured on any run of capturing the
data. Updated examples in man page. Change the default output path so that the
node name is not in the path. In addition I change how the clusternode name is
parsed.

Signed-off-by: Shane Bradley <sbradley@redhat.com>