fsck: Clean up pass1 inode iteration code
The original version of this patch, including readahead for the
inodes only appears to deliver a performance improvement in some
particular cases, and slows down other cases. So this patch is
fairly similar in that it includes the clean ups from the previous
attempt, however the readahead has been left out for the time
being until the problem is better understood, and can be fixed.

This clean up still has the advantage of better code structure
around the main loop, and in addition it also reduced the number
of read requests which are generated too.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>