fsck.gfs2: Don't flag GFS1 non-dinode blocks as duplicates
Before this patch, fsck.gfs2 could get into problems when processing
a GFS1 file system. The issue goes back to the fact that all GFS1
metadata is marked as "Meta" in the bitmap, whereas that bitmap
designation is reserved for dinodes in GFS2. For example, take a
GFS1 file of height 2, which looks like this:

Block
------
0x1234 dinode
0x1235 |----> indirect meta
0x1236        |---->data at offset 0 of the file

Before this patch, fsck.gfs2 would:

1. Encounter the dinode at 0x1234 and mark it as "dinode" in the
   blockmap.
2. Process its metadata, see block 0x1235, mark it as "indirect meta"
   in the blockmap.
3. Process the metadata's data, see block 0x1236, mark it as "data".
4. When it's done with the dinode, it moves on to the next dinode.
   But since GFS1 doesn't distinguish dinodes from other metadata,
   the next block in the bitmap that has that designation is block
   0x1235.
5. Since block 0x1235 was previously marked "indirect meta" pass1
   gets confused and thinks the block is a duplicate reference,
   and it's invalid as a dinode. This is a non-problem that's
   treated as a problem, and it makes bad decisions based on it,
   deleting what it perceives to be corruption.

This patch adds special checks for this problem and assumes the block
is just normal GFS1 non-dinode metadata.