gfs2-utils: Build system fixes
Previously we were disabling the whole test suite when check isn't
installed but tool_tests.sh doesn't require it so the conditional has
been moved into tests/Makefile.am so that tool_tests.sh is always
enabled.

Also fixes a problem where AC_PROG_YACC allowed the build to continue
when bison was not installed but byacc was. We use bison-specific
features so a specific check for bison has been added.

Signed-off-by: Andrew Price <anprice@redhat.com>