gfs2l: Add options to print block types and fields
This adds a -T option which prints out the list of possible meta header
types to be used in gfs2l scripts. It also adds -F <type> which prints
a list of fields in a block type given by <type>. The fields are
prefixed by their offset in the structure.

Signed-off-by: Andrew Price <anprice@redhat.com>