mkfs.gfs2: Create new resource groups on-demand
Adds a structure to encapsulate the state of resource group creation so
that the new rg_append() function can be called at any time. rg_append()
takes the number of data blocks required as a parameter so that future
enhancements can request a given number of free data blocks.

place_rgrps() now becomes a simple loop to call rg_append() and
writerg() until the end of the device.

Signed-off-by: Andrew Price <anprice@redhat.com>